+++ ibrcommon/ibrcommon/net/socket.h
@@ -215,8 +215,8 @@ namespace ibrcommon {
 		virtual void up() throw (socket_exception);
 		virtual void down() throw (socket_exception);
 
-		int send(const char *data, size_t len, int flags = 0) throw (socket_error);
-		int recv(char *data, size_t len, int flags = 0) throw (socket_error);
+		ssize_t send(const char *data, size_t len, int flags = 0) throw (socket_error);
+		ssize_t recv(char *data, size_t len, int flags = 0) throw (socket_error);
 
 		void set(CLIENT_OPTION opt, bool val) throw (socket_error);
